requirements.txt
absl-py==0.11.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
applaunchservices==0.2.1
appnope==0.1.0
argh==0.26.2
asgiref==3.3.1
asn1crypto==1.3.0
astor==0.8.1
astropy==4.0.1.post1
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.1
bkcharts==0.2
bleach==3.3.0
bokeh
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle
clyent==1.2.2
colorama==0.4.3
contextlib2==0.6.0.post1
contextvars==2.4
cryptography==3.3.2
cycler==0.10.0
Cython
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1594156306305/work
decorator==4.4.2
defusedxml==0.6.0
distributed
Django==3.1.13
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
fsspec==0.7.4
future==0.18.2
gast==0.2.2
gevent
glob2==0.7
gmpy2==2.0.8
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.33.2
h5py==2.10.0
HeapDict==1.0.1
html5lib
idna
imageio
imagesize==1.2.0
immutables
importlib-metadata
intervaltree
ipykernel
ipython
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi
Jinja2==2.11.3
joblib
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.10
jupyterlab-server
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.33.0+1.g022ab0f
locket==0.2.0
lxml
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.4
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpld3==0.5.2
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.7
networkx 
nose==1.3.7
notebook==6.4.1
numba==0.50.1
numexpr==2.7.1
numpy==1.17.0
numpydoc
olefile==0.46
opencv-python==4.4.0.46
openpyxl
opt-einsum==3.3.0
packaging==20.4
pandas
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle
pyflakes==2.2.0
Pygments==2.7.4
pylint
pyodbc===4.0.0-unsupported
pyOpenSSL
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-jsonrpc-server 
python-language-server
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.4
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole
QtPy==1.9.0
regex
requests
rope==0.17.0
Rtree==0.9.4
ruamel-yaml==0.15.87
scikit-image==0.16.2
scikit-learn
scipy==1.0.1
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport
spyder
spyder-kernels
SQLAlchemy
sqlparse==0.4.2
statsmodels==0.11.1
sympy
tables==3.6.1
tblib==1.6.0
tensorboard==1.15.0
tensorflow==2.5.1
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl
toml
toolz==0.10.0
tornado==6.0.4
tqdm
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions
ujson==1.35
unicodecsv==0.14.1
urllib3==1.26.5
watchdog
wcwidth
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.2.0
yapf
yfinance==0.1.55
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1